requirements.txt
Flask==1.1.1
requests==2.20.0
google-cloud-datastore==1.10.0
webapp2==2.5.2
google-cloud-storage==1.26.0
google-cloud-pubsub==1.3.0